заменил Float на Numeric для цены, исправил init_db.py

+++ app/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 from app.db.db import Base
 
@@ -18,7 +18,7 @@ class Book(Base):
     id = Column(Integer, primary_key=True)
     title = Column(String, nullable=False)
     description = Column(String)
-    price = Column(Float)
+    price = Column(Numeric(10, 2))
     url = Column(String)
     category_id = Column(Integer, ForeignKey("categories.id"))
